Added UI layer controller for MenuView
